uk-cca-app-api/src/test/java/uk/gov/cca/api/workflow/request/flow/admintermination/withdraw/handler/AdminTerminationWithdrawSaveActionHandlerTest.java
package uk.gov.cca.api.workflow.request.flow.admintermination.withdraw.handler;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import uk.gov.cca.api.workflow.request.core.domain.CcaRequestTaskActionType;
import uk.gov.cca.api.workflow.request.flow.admintermination.withdraw.domain.AdminTerminationWithdrawSaveRequestTaskActionPayload;
import uk.gov.cca.api.workflow.request.flow.admintermination.withdraw.service.AdminTerminationWithdrawService;
import uk.gov.netz.api.authorization.core.domain.AppUser;
import uk.gov.netz.api.workflow.request.core.domain.RequestTask;
import uk.gov.netz.api.workflow.request.core.service.RequestTaskService;

import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
import static org.mockito.Mockito.*;

@ExtendWith(MockitoExtension.class)
public class AdminTerminationWithdrawSaveActionHandlerTest {

    @InjectMocks
    private AdminTerminationWithdrawSaveActionHandler handler;

    @Mock
    private RequestTaskService requestTaskService;

    @Mock
    private AdminTerminationWithdrawService adminTerminationWithdrawService;

    @Test
    void process() {
        final AppUser appUser = AppUser.builder().build();
        final long requestTaskId = 1L;
        final String requestTaskActionType = CcaRequestTaskActionType.ADMIN_TERMINATION_WITHDRAW_SAVE_APPLICATION;
        final AdminTerminationWithdrawSaveRequestTaskActionPayload taskActionPayload =
                AdminTerminationWithdrawSaveRequestTaskActionPayload.builder().build();

        final RequestTask requestTask = RequestTask.builder().id(requestTaskId).build();

        when(requestTaskService.findTaskById(requestTaskId)).thenReturn(requestTask);

        // Invoke
        handler.process(requestTaskId, requestTaskActionType, appUser, taskActionPayload);

        // Verify
        verify(requestTaskService, times(1)).findTaskById(requestTaskId);
        verify(adminTerminationWithdrawService, times(1))
                .applySaveAction(taskActionPayload, requestTask);
    }

    @Test
    void getTypes() {
        assertThat(handler.getTypes())
                .containsExactly(CcaRequestTaskActionType.ADMIN_TERMINATION_WITHDRAW_SAVE_APPLICATION);
    }
}

